+++ App/Targets/AppInterface/Sources/AppModuleInterface.swift
@@ -22,11 +22,6 @@ public protocol AppModuleInterface {
     func subscribeMarketingFCMTopic(completion: @escaping ((Error?) -> Void))
     func unsubscribeMarketingFCMTopic(completion: @escaping ((Error?) -> Void))
     
-    func presentMailComposeViewController(
-        nickname: String,
-        targetViewController: UIViewController
-    )
-    
     /// GA
     func sendPageView(screenName: String, type: AnyObject.Type)
     func sendEvent(name: String, parameters: [String: Any]?)
