+++ Modules/Feature/MyPage/Targets/MyPage/Sources/Domains/Qna/QnaViewController.swift
@@ -54,10 +54,7 @@ public final class QnaViewController: BaseViewController {
         case .pushFAQ:
             pushFAQ()
         case .presentMail(let nickname):
-            Environment.appModuleInterface.presentMailComposeViewController(
-                nickname: nickname,
-                targetViewController: self
-            )
+            presentMailComposer(nickname: nickname)
         }
     }
     
@@ -65,6 +62,25 @@ public final class QnaViewController: BaseViewController {
         let viewController = FaqViewController()
         
         navigationController?.pushViewController(viewController, animated: true)
+    }
+    
+    private func presentMailComposer(nickname: String) {
+        guard MFMailComposeViewController.canSendMail() else {
+            return
+        }
+        
+        let iosVersion = UIDevice.current.systemVersion
+        let appVersion = Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String ?? ""
+        let deviceModel = Device.current
+        
+        let composer = MFMailComposeViewController()
+        composer.mailComposeDelegate = self
+        composer.setToRecipients(["3dollarinmypocket@gmail.com"])
+        composer.setSubject("가슴속 3천원 문의")
+        composer.setMessageBody("\n\n\n\n----------\n닉네임: \(nickname)\n앱 버전: \(appVersion)\nOS: ios \(iosVersion)\n디바이스: \(deviceModel)", isHTML: false)
+        
+        
+        present(composer, animated: true)
     }
 }
